Move runtime apis to common
