+++ runtime/common/src/lib.rs
@@ -1,4 +1,5 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
+pub mod runtime_apis;
 pub mod constants;
 pub mod types;
